feat(): replace old personal info tab with new one
