--- client/src/components/App/Content/InvestigationForm/TabManagement/NewPersonalInfoTab/PersonalInfoTabInterfaces.ts
@@ -1,16 +1,0 @@
-import PersonalInfoDataContextFields from 'models/enums/PersonalInfoDataContextFields';
-
-export interface PersonalInfoTabState {
-    [PersonalInfoDataContextFields.PHONE_NUMBER]: string;
-    [PersonalInfoDataContextFields.ADDITIONAL_PHONE_NUMBER]: string;
-    [PersonalInfoDataContextFields.CONTACT_PHONE_NUMBER]: string;
-    [PersonalInfoDataContextFields.CONTACT_INFO]: string;
-    [PersonalInfoDataContextFields.INSURANCE_COMPANY]: string;
-    [PersonalInfoDataContextFields.RELEVANT_OCCUPATION]: string;
-    [PersonalInfoDataContextFields.INSTITUTION_NAME]: string;
-    [PersonalInfoDataContextFields.EDUCATION_OCCUPATION_CITY]?: string;
-    [PersonalInfoDataContextFields.ROLE]?: number;
-    [PersonalInfoDataContextFields.EDUCATION_GRADE]?: number;
-    [PersonalInfoDataContextFields.EDUCATION_CLASS_NUMBER]?: number;
-    [PersonalInfoDataContextFields.OTHER_OCCUPATION_EXTRA_INFO]?: string;
-}
